add assignment2
Assignment Overview

In this assignment you will update dishdetail.html and app.js.
At the end of this assignment, you should have completed the
 following tasks to update the page:

Provide a form to enable users to submit their comments
Validate the information entered by the users in the form
Show a real-time preview of the comment on the page.